requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.5
grpcio==1.67.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
jiter==0.8.0
milvus-lite==2.4.10
numpy==2.0.2
openai==1.56.1
outcome==1.3.0.post0
pandas==2.2.3
protobuf==5.29.0
pydantic==2.10.3
pydantic_core==2.27.1
pymilvus==2.5.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
selenium==4.27.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.41.3
tqdm==4.67.1
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.32.1
websocket-client==1.8.0
wsproto==1.2.0
